+++ R/api_exports.R
@@ -12,9 +12,8 @@
 #' Can be a plotly/ggplot2 object or a \link{data.frame}.
 #' @param filename character vector naming file(s). If \code{x} is a plot,
 #' can be a vector of length 2 naming both the plot AND the underlying grid.
-#' If a filename matches a file that already exists, you will be prompted with 
-#' an option to overwrite it (if called non-interactively, the file will be 
-#' overwritten automatically, unless it is a folder).
+#' @param fileopt character string describing whether to "overwrite" existing 
+#' files or ensure "new" file(s) are always created.
 #' @param sharing If 'public', anyone can view this graph. It will appear in 
 #' your profile and can appear in search engines. You do not need to be
 #' logged in to Plotly to view this chart.
@@ -113,13 +112,12 @@
 
 #' @rdname api
 #' @export
-api_create <- function(x = last_plot(), filename = NULL,
+api_create <- function(x = last_plot(), filename = NULL, 
+                       fileopt = c("overwrite", "new"),
                        sharing = c("public", "private", "secret"), ...) {
-  dots <- list(...)
-  if (!is.null(dots[["fileopt"]])) {
-    warning("The fileopt argument is no longer supported", call. = FALSE)
-  }
-  UseMethod("api_create", x)
+  fileopt <- match.arg(fileopt, c("overwrite", "new"))
+  sharing <- match.arg(sharing, c("public", "private", "secret"))
+  UseMethod("api_create")
 }